Create SharePoint 2013 Use HTTPs in a remote app and use OAuth for callbacks
Contents
SharePoint 2013: Use HTTPs in a remote app	2
Description of the sample	2
Prerequisites	2
Key components of the sample	2
Configure the sample	3
Build the sample	3
Run and test the sample	3
Example	3
Troubleshooting	4
Change log	5
Related content	5

 
SharePoint 2013: Use HTTPs in a remote app
This documentation is preliminary and is subject to change.
Summary:  This code sample uses HTTPS in a basic provider-hosted app for SharePoint. It also uses a combination of the SharePoint REST interface and the client object model (CSOM) to read information about a SharePoint 2013 Preview site.
This sample demonstrates a basic provider-hosted app for SharePoint that uses HTTPS. It also uses a combination of the SharePoint REST interface and the SharePoint client object model (CSOM) to read and display information about the parent SharePoint 2013 Preview site on which the app is installed.
Description of the sample
The code that uses the REST API and the CSOM is located in the Home.aspx.cs file of the RemoteAppUsingHttpsWeb project. The sample uses HTTPS. It reads and displays OAuth tokens. It also reads data that conforms with the OData protocol from the REST endpoints where the basic SharePoint entities, such as users and site information, are exposed. It also reads data using the SharePoint CSOM.
 	Note
For more information about the SharePoint REST APIs, see Programming using the SharePoint 2013 REST service. For more information about working with JSON, Atom, and OData, see OData: JavaScript Object Notation (JSON) Format and OData: AtomPub Format.
Prerequisites
This sample requires the following;
•	You must know how to create a provider-hosted app for SharePoint. For more information about how to create a provider-hosted app for SharePoint, see How to: Create a basic provider-hosted app for SharePoint.
•	You must know how to use a certificate.
•	A SharePoint 2013 Preview development environment that is configured for app isolation and OAuth.
•	Visual Studio 2012 and SharePoint development tools in Visual Studio 2012.
•	Basic familiarity with REST web services.
Key components of the sample
The sample app contains the following:
•	The RemoteAppUsingHttps project, which contains the AppManifest.xml file.
•	The RemoteAppUsingHttpsWeb project, which contains:
•	The Home.aspx file, which contains the HTML and ASP.NET controls for the user interface of the app.
•	The Home.aspx.cs file, which contains the C# code that uses the CSOM and REST to read and write data.
•	The web.config file.
Configure the sample
To configure the basic data operations sample app:
•	Update the SiteUrl property of the solution with the URL of the home page of your SharePoint 2013 Preview site.
Build the sample
Press F5 to build and deploy the app.
Run and test the sample
1.	Press F5 to build and deploy the app.
2.	On the consent page, choose Trust It to grant permissions to the app.
Example
Figure 1 shows the sample URL you see when you launch the app. 
Figure 1. Start page URL of the app after you launch it
  Figure 2 shows an example of the kinds of information that this sample app reads and displays.
Figure 2. View of the populated page
  
Troubleshooting
The following table lists common configuration and environment errors that prevent the sample from running or deploying properly and how to solve them.
 
Problem	Solution
Visual Studio does not open the browser after you press the F5 key.	Set the app for SharePoint project as the startup project.
HTTP error 405 Method not allowed.	Locate the applicationhost.config file in %userprofile%\Documents\IISExpress\config.
Locate the handler entry for StaticFile, and add the verbs GET, HEAD, POST, DEBUG, and TRACE.
 
Change log
 
Version	Date
First version	July 16, 2012
 
Related content
•	How to: Create a basic provider-hosted app for SharePoint
•	App permissions in SharePoint 2013
•	OAuth authentication and authorization flow for cloud-hosted apps
•	Tips and FAQs: OAuth and remote apps for SharePoint 2013
•	App authorization policy types
•	SharePoint 2013 development overview
•	Get started developing apps for SharePoint
•	How to: Create a basic SharePoint-hosted app
•	Authorization and authentication for apps in SharePoint 2013
•	Critical aspects of the app for SharePoint 2013 architecture and development landscape
•	Apps for SharePoint compared with SharePoint solutions
•	Critical aspects of the app for SharePoint 2013 architecture and development landscape
 
 